........
  r253536 | russell | 2010-03-20 06:33:30 -0500 (Sat, 20 Mar 2010) | 4 lines
  
  Use SHRT_MAX instead of MAXSHORT.
  
  These changes fix build issues I had with this module on FreeBSD.
........
  r253537 | russell | 2010-03-20 06:39:39 -0500 (Sat, 20 Mar 2010) | 2 lines
  
  Resolve a compiler warning on FreeBSD.
........
  r253538 | russell | 2010-03-20 06:43:08 -0500 (Sat, 20 Mar 2010) | 2 lines
  
  Resolve compiler warnings on FreeBSD.
........
  r253540 | russell | 2010-03-20 07:03:07 -0500 (Sat, 20 Mar 2010) | 2 lines
  
  Resolve more compiler warnings on FreeBSD.
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.6.0@253626 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_dial.c
@@ -1274,12 +1274,12 @@ static void end_bridge_callback(void *data)
 
 	ast_channel_lock(chan);
 	if (chan->cdr->answer.tv_sec) {
-		snprintf(buf, sizeof(buf), "%ld", end - chan->cdr->answer.tv_sec);
+		snprintf(buf, sizeof(buf), "%ld", (long) end - chan->cdr->answer.tv_sec);
 		pbx_builtin_setvar_helper(chan, "ANSWEREDTIME", buf);
 	}
 
 	if (chan->cdr->start.tv_sec) {
-		snprintf(buf, sizeof(buf), "%ld", end - chan->cdr->start.tv_sec);
+		snprintf(buf, sizeof(buf), "%ld", (long) end - chan->cdr->start.tv_sec);
 		pbx_builtin_setvar_helper(chan, "DIALEDTIME", buf);
 	}
 	ast_channel_unlock(chan);
